DS-687 | @ericbakenhus | Add generated fields for Algolia config (#860)
* DS-687: Add generated fields

* DS-687: Add geo location data

* DS-687: Check rendered description
